+++ exclusiveCars/src/main/java/com/fmi/exclusiveCars/services/NewsService.java
@@ -1,5 +1,6 @@
 package com.fmi.exclusiveCars.services;
 
+import com.fmi.exclusiveCars.dto.NewsDto;
 import com.fmi.exclusiveCars.model.News;
 import com.fmi.exclusiveCars.repository.NewsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,9 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -20,42 +24,72 @@ public class NewsService {
         this.newsRepository = newsRepository;
     }
 
+    public ResponseEntity<?> getAllNews() {
+        Collection<News> news = newsRepository.findAll();
+        if(news.isEmpty()) {
+            return new ResponseEntity<>("There are no news published yet!", HttpStatus.OK);
+        }
+
+        List<News> newsList = new ArrayList<>(news);
+        return new ResponseEntity<>(newsList, HttpStatus.OK);
+    }
+
     public ResponseEntity<?> getNews(Long id) {
         Optional<News> pieceOfNews = newsRepository.findById(id);
 
         if(pieceOfNews.isPresent()) {
             return new ResponseEntity<>(pieceOfNews.get(), HttpStatus.OK);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>("The piece of news you asked for doesn't exist!", HttpStatus.NOT_FOUND);
     }
 
-    public Long addNews(News pieceOfNews) {
-        pieceOfNews.setDate(LocalDate.now());
-        pieceOfNews.setHour(LocalTime.now());
-        News newPieceOfNews = newsRepository.save(pieceOfNews);
+    public ResponseEntity<?> addNews(NewsDto pieceOfNews) {
+        Optional<News> news = newsRepository.findByTitle(pieceOfNews.getTitle());
+        if(news.isPresent()) {
+            return new ResponseEntity<>("There is already a piece of news with the same title!", HttpStatus.BAD_REQUEST);
+        }
 
-        return newPieceOfNews.getId();
+        News newsToAdd = News.builder()
+                .title(pieceOfNews.getTitle())
+                .content(pieceOfNews.getContent())
+                .date(LocalDate.now())
+                .hour(LocalTime.now())
+                .build();
+        newsRepository.save(newsToAdd);
+
+        return new ResponseEntity<>("The news article was successfully added!", HttpStatus.OK);
     }
 
-    public ResponseEntity<?> editNews(Long id, News news) {
+    public ResponseEntity<?> editNews(Long id, NewsDto news) {
         Optional<News> pieceOfNews = newsRepository.findById(id);
 
+        Optional<News> newsByTitle = newsRepository.findByTitle(news.getTitle());
+
         if(pieceOfNews.isPresent()) {
+            if(newsByTitle.isPresent() && newsByTitle.get() != pieceOfNews.get()) {
+                return new ResponseEntity<>("There is already a piece of news with the same title!", HttpStatus.BAD_REQUEST);
+            }
+
             News currentNews = pieceOfNews.get();
             currentNews.setTitle(news.getTitle());
             currentNews.setContent(news.getContent());
-            currentNews.setDate(news.getDate());
-            currentNews.setHour(news.getHour());
+            currentNews.setDate(LocalDate.now());
+            currentNews.setHour(LocalTime.now());
             newsRepository.save(currentNews);
 
-            return new ResponseEntity<>(currentNews, HttpStatus.OK);
+            return new ResponseEntity<>("The news article was edited successfully!", HttpStatus.OK);
         }
 
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>("The news article you have requested to edit doesn't exist!", HttpStatus.NOT_FOUND);
     }
 
-    public ResponseEntity<HttpStatus> deleteNews(Long id) {
+    public ResponseEntity<?> deleteNews(Long id) {
+        Optional<News> pieceOfNews = newsRepository.findById(id);
+
+        if(pieceOfNews.isEmpty()) {
+            return new ResponseEntity<>("The news article you have requested to delete doesn't exist!", HttpStatus.NOT_FOUND);
+        }
         newsRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>("The piece of news was successfully deleted!", HttpStatus.OK);
     }
 }
