+++ exclusiveCars/src/main/java/com/fmi/exclusiveCars/services/RentalCenterService.java
@@ -1,5 +1,6 @@
 package com.fmi.exclusiveCars.services;
 
+import com.fmi.exclusiveCars.dto.RentalCenterDto;
 import com.fmi.exclusiveCars.model.RentalCenter;
 import com.fmi.exclusiveCars.repository.RentalCenterRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +8,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -18,6 +22,16 @@ public class RentalCenterService {
         this.rentalCenterRepository = rentalCenterRepository;
     }
 
+    public ResponseEntity<?> getAllRentalCenters() {
+        Collection<RentalCenter> rentalCenters = rentalCenterRepository.findAll();
+        if(rentalCenters.isEmpty()) {
+            return new ResponseEntity<>("There are no rental centers listed yet!", HttpStatus.OK);
+        }
+
+        List<RentalCenter> rentalCenterList = new ArrayList<>(rentalCenters);
+        return new ResponseEntity<>(rentalCenterList, HttpStatus.OK);
+    }
+
     public ResponseEntity<?> getRentalCenter(Long id) {
         Optional<RentalCenter> rentalCenter = rentalCenterRepository.findById(id);
 
@@ -26,34 +40,64 @@ public class RentalCenterService {
             return new ResponseEntity<>(currentRentalCenter, HttpStatus.OK);
         }
 
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>("The rental center you asked for doesn't exist!", HttpStatus.NOT_FOUND);
     }
 
-    public Long addRentalCenter(RentalCenter rentalCenter) {
-        rentalCenterRepository.save(rentalCenter);
-        return rentalCenter.getId();
-    }
-
-    public ResponseEntity<?> editRentalCenter(Long id, RentalCenter rentalCenter) {
-        Optional<RentalCenter> actualRentalCenter = rentalCenterRepository.findById(id);
-
-        if(actualRentalCenter.isPresent()) {
-            RentalCenter currentRentalCenter = actualRentalCenter.get();
-            currentRentalCenter.setName(rentalCenter.getName());
-            currentRentalCenter.setCity(rentalCenter.getCity());
-            currentRentalCenter.setAddress(rentalCenter.getAddress());
-            currentRentalCenter.setEmail(rentalCenter.getEmail());
-            currentRentalCenter.setPhone(rentalCenter.getPhone());
-
-            rentalCenterRepository.save(currentRentalCenter);
-            return new ResponseEntity<>(currentRentalCenter, HttpStatus.OK);
+    public ResponseEntity<?> addRentalCenter(RentalCenterDto rentalCenterDto) {
+        Optional<RentalCenter> rentalCenterByName = rentalCenterRepository.findByName(rentalCenterDto.getName());
+        Optional<RentalCenter> rentalCenterByEmail = rentalCenterRepository.findByEmail(rentalCenterDto.getEmail());
+        Optional<RentalCenter> rentalCenterByPhone = rentalCenterRepository.findByPhone(rentalCenterDto.getPhone());
+        if(rentalCenterByName.isPresent() || rentalCenterByEmail.isPresent() || rentalCenterByPhone.isPresent()) {
+            return new ResponseEntity<>("There is already a rental center with this information!", HttpStatus.BAD_REQUEST);
         }
 
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        RentalCenter rentalCenterToAdd = RentalCenter.builder()
+                .name(rentalCenterDto.getName())
+                .city(rentalCenterDto.getCity())
+                .address(rentalCenterDto.getAddress())
+                .email(rentalCenterDto.getEmail())
+                .phone(rentalCenterDto.getPhone())
+                .build();
+        rentalCenterRepository.save(rentalCenterToAdd);
+
+        return new ResponseEntity<>("The rental center was successfully added!", HttpStatus.OK);
     }
 
-    public ResponseEntity<HttpStatus> deleteRentalCenter(Long id) {
+    public ResponseEntity<?> editRentalCenter(Long id, RentalCenterDto rentalCenterDto) {
+        Optional<RentalCenter> actualRentalCenter = rentalCenterRepository.findById(id);
+
+        Optional<RentalCenter> rentalCenterByName = rentalCenterRepository.findByName(rentalCenterDto.getName());
+        Optional<RentalCenter> rentalCenterByEmail = rentalCenterRepository.findByEmail(rentalCenterDto.getEmail());
+        Optional<RentalCenter> rentalCenterByPhone = rentalCenterRepository.findByPhone(rentalCenterDto.getPhone());
+
+        if(actualRentalCenter.isPresent()) {
+            if((rentalCenterByName.isPresent() && rentalCenterByName.get() != actualRentalCenter.get())
+                    || (rentalCenterByEmail.isPresent() && rentalCenterByEmail.get() != actualRentalCenter.get())
+                    || (rentalCenterByPhone.isPresent() && rentalCenterByPhone.get() != actualRentalCenter.get())) {
+                return new ResponseEntity<>("There is already a rental center with this information!", HttpStatus.BAD_REQUEST);
+            }
+
+            RentalCenter currentRentalCenter = actualRentalCenter.get();
+            currentRentalCenter.setName(rentalCenterDto.getName());
+            currentRentalCenter.setCity(rentalCenterDto.getCity());
+            currentRentalCenter.setAddress(rentalCenterDto.getAddress());
+            currentRentalCenter.setEmail(rentalCenterDto.getEmail());
+            currentRentalCenter.setPhone(rentalCenterDto.getPhone());
+
+            rentalCenterRepository.save(currentRentalCenter);
+            return new ResponseEntity<>("The rental center was successfully edited!", HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>("The rental center you have requested to edit doesn't exist!", HttpStatus.NOT_FOUND);
+    }
+
+    public ResponseEntity<?> deleteRentalCenter(Long id) {
+        Optional<RentalCenter> rentalCenter = rentalCenterRepository.findById(id);
+
+        if(rentalCenter.isEmpty()) {
+            return new ResponseEntity<>("The rental center you requested to delete doesn't exist!", HttpStatus.NOT_FOUND);
+        }
         rentalCenterRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>("The rental center was successfully deleted!", HttpStatus.OK);
     }
 }
